+++ src/AzureIoTHub.Portal.Infrastructure/Startup/AWSServiceCollectionExtension.cs
@@ -1,0 +1,24 @@
+// Copyright (c) CGI France. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace AzureIoTHub.Portal.Infrastructure.Startup
+{
+    using AzureIoTHub.Portal.Infrastructure.ServicesHealthCheck;
+    using Microsoft.Extensions.DependencyInjection;
+
+    public static class AWSServiceCollectionExtension
+    {
+        public static IServiceCollection AddAWSInfrastructureLayer(this IServiceCollection services)
+        {
+            return services.ConfigureHealthCheck();
+        }
+        private static IServiceCollection ConfigureHealthCheck(this IServiceCollection services)
+        {
+            _ = services.AddHealthChecks()
+                        .AddDbContextCheck<PortalDbContext>()
+                        .AddCheck<DatabaseHealthCheck>("databaseHealthCheck");
+
+            return services;
+        }
+    }
+}
